Refine AI chat permission gating
