+++ server/tests/capabilities.test.ts
@@ -3,14 +3,24 @@ import { UserRole } from '../types/main';
 import { resolveEffectiveCapabilities } from '../authz/capabilities';
 
 describe('capability resolution', () => {
+  it('keeps AI chat denied for users and allowed for admins by default', () => {
+    const userCapabilities = resolveEffectiveCapabilities({ role: UserRole.USER });
+    const moderatorCapabilities = resolveEffectiveCapabilities({ role: UserRole.MODERATOR });
+    const adminCapabilities = resolveEffectiveCapabilities({ role: UserRole.ADMIN });
+
+    expect(userCapabilities['ai.chat'].allowed).toBe(false);
+    expect(moderatorCapabilities['ai.chat'].allowed).toBe(false);
+    expect(adminCapabilities['ai.chat'].allowed).toBe(true);
+  });
+
   it('uses user overrides before role policies', () => {
     const capabilities = resolveEffectiveCapabilities({
       role: UserRole.USER,
-      rolePolicies: { 'ai.site.chat': 'deny' },
-      userOverrides: { 'ai.site.chat': 'allow' },
+      rolePolicies: { 'ai.chat': 'deny' },
+      userOverrides: { 'ai.chat': 'allow' },
     });
 
-    expect(capabilities['ai.site.chat']).toEqual({
+    expect(capabilities['ai.chat']).toEqual({
       allowed: true,
       source: 'user_override',
       role: UserRole.USER,
@@ -36,7 +46,7 @@ describe('capability resolution', () => {
   it('does not allow policies or overrides to reduce super admin permissions', () => {
     const capabilities = resolveEffectiveCapabilities({
       role: UserRole.SUPER_ADMIN,
-      rolePolicies: { 'ai.site.chat': 'deny' },
+      rolePolicies: { 'ai.chat': 'deny' },
       userOverrides: { 'attachment.upload': 'deny' },
     });
