+++ server/utils/dbMethods/ai/jobs.ts
@@ -1,0 +1,15 @@
+export { ensurePgvectorReady, getPgvectorStatus } from './vector';
+export {
+  deleteEmbeddingsForOwner,
+  deleteEmbeddingsForSource,
+  enqueueBackfillEmbeddingJobs,
+  enqueueBackfillEmbeddingJobsForOwner,
+  enqueueEmbeddingJob,
+  getEmbeddingJobStats,
+} from './embeddingQueue';
+export {
+  clearAllEmbeddings,
+  processPendingEmbeddingJobs,
+  retryFailedEmbeddingJobs,
+  setIndexingPaused,
+} from './embeddingWorker';
