+++ server/utils/dbMethods/ai/search.ts
@@ -1,0 +1,2 @@
+export { semanticSearch } from './semanticSearch';
+export { searchMemoryWithFallback, textSearchMemory } from './textSearch';
